refactoring: - css 파일 assets 폴더로 분리 - axiosConfig 파일 사용
