--- src/components/ApplyStatusForStaff/ApplyStatus.css
@@ -1,44 +1,0 @@
-.party-div {
-  margin-top: 3%;
-}
-
-.party-h1 {
-  display: inline;
-  color: #6AC9C6;
-}
-
-.apply-h4 {
-  display: inline;
-}
-
-h3 {
-  margin-bottom: 0;
-}
-
-.de-step-div button {
-  width: fit-content;
-  padding: 10px;
-  margin-top: 0%;
-  margin-right: 7px;
-  margin-bottom: 5px;
-  font-size: medium;
-  border: 0px;
-}
-
-.department-h4 {
-  margin-top: 5%;
-}
-
-.applicant-p {
-  display: inline;
-  color: #6AC9C6;;
-  font-weight: 700;
-  font-size: 1.1em;
-}
-
-.competition-p {
-  color: #6AC9C6;;
-  font-weight: 600;
-  font-size: 1.0em;
-  margin-top: 8px;
-}
