--- src/components/ApplyStatusForStaff/ApplyStatusForStaff.css
@@ -1,5 +1,0 @@
-.wrap_div {
-    width: 60%;
-    margin: auto;
-    margin-top: 3%;
-}
