--- src/components/ApplyStatusForStaff/Buttons.css
@@ -1,14 +1,0 @@
-.button-div {
-  margin-top: 40px;
-  padding-left: 20px;
-  padding-right: 20px;
-}
-
-.buttons {
-  width: 49%;
-  margin-bottom: 10px;
-  padding: 10px;
-  border: 0px;
-  background-color: #C4C4C4;
-  font-weight: 700;
-}
