--- src/components/GradingStatusForStaff/GradingTable.css
@@ -1,37 +1,0 @@
-.status-h4 {
-  padding-bottom: 7px;
-}
-
-table {
-  width: 100%; 
-  padding-top: 7px; 
-  text-align: center; 
-  background: #f2f2f2;
-}
-
-tr.thead td {
-  padding-bottom: 0px;
-}
-
-td {
-  padding-bottom: 10px;
-}
-
-.padding {
-  height: 10px;
-}
-
-.full-button {
-    width: 19px;
-    height: 19px;
-    background-color: #6AC9C6;
-    border: 0;
-    border-radius: 100%;
-}
-
-.empty-button {
-    width: 20px;
-    height: 20px;
-    border: 2px solid #6AC9C6;
-    border-radius: 100%;
-}
