--- src/components/GradingStatusForStaff/Guideline.css
@@ -1,19 +1,0 @@
-.guide-div {
-  margin-top: 5%;
-  background-color: #C4C4C4;
-  height: 80px;
-  width: fit-content;
-  margin: auto;
-  margin-top: 5%;
-  margin-bottom: 7%;
-  padding-top: 8px;
-  padding-left: 2%;
-  padding-right: 2%;
-  border-radius: 10px;
-}
-
-p {
-  padding-bottom: 1%;
-  font-size: 0.9em;
-  font-weight: bold;
-}
