--- src/components/GradingStatusForStaff/Result.css
@@ -1,7 +1,0 @@
-.result-div {
-  margin-top: 5%;
-}
-
-.result-h5 {
-  display: inline;
-}
