--- src/components/main/Main.css
@@ -1,8 +1,0 @@
-.wrap_div {
-  width: 60%;
-  margin: auto;
-}
-
-h4 {
-  height: 10px;
-}
