--- src/components/main/RecruitmentBlock.css
@@ -1,89 +1,0 @@
-.odd1 {
-  width: 47%;
-  height: 150px;
-  margin-bottom: 30px;
-  float: left;
-  padding-bottom: 10px;
-  background-color: #EDEDED;
-  box-shadow: 4px 2px 2px #dadada;
-}
-
-.odd0 {
-  width: 47%;
-  height: 150px;
-  margin-bottom: 30px;
-  float: right;
-  padding-bottom: 10px;
-  background-color: #EDEDED;
-  box-shadow: 4px 2px 2px #dadada;
-}
-
-.party-div {
-  margin-bottom: 10px;
-  padding: 10px;
-  padding-left: 20px;
-}
-
-.party-name {
-  margin-top: 0;
-  height: 7px;
-  display: inline;
-}
-
-.department-p1 {
-  display: inline;
-  font-size: 0.87em;
-}
-
-.department-p {
-  display: inline;
-  font-size: 0.87em;
-  font-weight: 900;
-  color: #6AC9C6;;
-}
-
-.availability-div {
-  width: 20%;
-  height: 22%;
-  float: right;
-  margin-top: 0;
-  background-color: #84D5D5;
-  text-align: center;
-  line-height: 15%;
-}
-
-.availability-div p {
-  color: white;
-}
-
-.date-div {
-  font-size: small;
-  padding-left: 20px;
-}
-
-p {
-  height: 5px;
-}
-
-.competition-div {
-  color: #84D5D5;
-  margin-top: 7px;
-}
-
-.competition-props {
-  height: fit-content;
-  color: #84D5D5;
-  font-size: 0.78em;
-  font-weight: 550;
-  margin: 0;
-}
-
-.step-p {
-  display: inline;
-  font-weight: 550;
-}
-
-.date-p {
-  display: inline;
-  font-size: 0.87em;
-}
