+++ src/components/main/RecruitmentBlock.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import "./RecruitmentBlock.css";
+import "../../assets/Main/RecruitmentBlock.css";
+import "../../assets/Main/Main.css";
 
 function RecruitmentBlock(props) {
   return (
