+++ src/pages/Home.js
@@ -1,11 +1,11 @@
 import React, { useEffect } from 'react';
-import axios from 'axios';
+import instance from '../../utils/axiosConfig';
 
 function Home() {
   const authorization = window.localStorage.getItem('authorization');
   const getData = async () => {
     try {
-      const res = await axios.get('http://localhost:8787/result', {
+      const res = await instance.get('http://localhost:8787/result', {
         headers: {
           authorization: authorization,
         },
